Add support for durable AMQP queues
The collectd amqp plugin allows setting messages as persistent but is hardcoded to create or use transient queues that are set to auto_delete.  This commit adds support for QueueDurable and QueueAutoDelete boolean values in the config file and changes the queue creation as appropriate.

If the values are not set in the config file, QueueDurable defaults to false and QueueAutoDelete defaults to true.

Tested with RabbitMQ 3.3.1