Remove spammy debug message from putval.
Other messages are printed to the socket in this section of the code,
but they are all error messages and therefore helpful.  This message
appears to be more of a debug message indicating that things are working
normally.  As people have reported in #664, in production scenarios it
floods logs and is spammy and unhelpful.