Fix overquoting in htonll tests
Similar to 4def66d6820b46e9044c4e2a0e30b043b330e8d2, the overquoting
here is not completely undone, leaving spurious `[` and `]` characters
in the C source, which leads to a compiler failure and the determination
that HAVE_HTONLL=no even if, as on OS X Yosemite, it is defined.

There are other instances of [[[ ]]] in configure.ac, but as always
touching things that aren't known to be broken seems dangerous!