add MemAvailable support to the memory plugin
A new metric called 'MemAvailable' was added to /proc/meminfo in the 3.14 kernel, which is an "estimate of how much memory is available for starting new applications, without swapping. Calculated from MemFree, SReclaimable, the size of the file LRU lists, and the low watermarks in each zone."

This patch added support for this metric to the memory plugin.