Add a write_log output plugin which logs values.
I find this to be useful when developping new input plugins,
it allows creating a very simple configuration and combined
with `-f` can be used for a very simple workflow.

I'm proposing this, since I think it could be useful for
people wanting to debug their config on local agents as well.