Turbostat: reorganise metric names
The `plugin_instance` now always is the hardware element
(cpuNN/coreNN/pkgNN) the data belongs to. Any optional identifier string
is moved to the `type_instance`.

Also remove the "_W" unit suffix as the "power" `type` implies we're
dealing with Watts.

The idea is to have the naming scheme in line with the way most other
plugins usually organise data.