disk plugin: Fix compatibility for Mac OS X 10.10.
Here’s a fixed version of the disk plugin that makes it work under Mac
OSX 10.10, and hopefully previous releases. The original version was
looking for the disk name in one dictionary, but it was actually in
another. I suspect at some point, Apple moved the disk name property,
but I don’t have any previous releases to check.

I changed the code to look for the disk name in both dictionaries, so
hopefully it’s backward compatible.

Signed-off-by: Florian Forster <octo@collectd.org>