src/daemon/plugin.c: delay first plugin read
This commit adds an optional configuration parameter SyncNodesReads to
delay the first plugin read to the beginning of the next read interval.
This way, the plugin reads are more or less synchronized accross
multiple nodes running collectd with the same read interval. This helps
comparing metrics shared by multiple nodes, such as network bandwidth
and so on.

Please note that it only works if the system clocks of these nodes are
properly synchronized using NTP (or similar) but this is generally (and
hopefully) the case on large server infrastructures.

The default value of this setting is false. It means that unless the
parameter is explicitely set to true, there is no change to previous
behaviour: the first plugin read is scheduled as soon as possible on
collectd start.