utils_buffer: Add buffer utilities.
This is a data structure describing an in-memory buffer, which can be
written to in a manner similar to a file with functions equivalent to
putc(), puts(), printf() and friends.

The buffer can be static or can grow dynamically up to a configurable
upper limit.

The implementation tries to be as safe as possible. All operations
perform bounds checking and handle errors. The presence of a trailing
NUL byte is also ensured.