utils_curl: Add CURL utilities.
This provides an easy-to-use abstraction layer on top of a CURLM
handle. It integrates with the "utils_iomux" code to facilitate
simultaneous processing of an arbitrary number of CURL requests within
a single thread.

The code is designed to be thread-safe so that multiple threads can
share a single CURLM instance.