utils_hashtable: Add generic hash table implemenation.
This is an implementation of a hash table that is hopefully generic
and reusable enough to be generally useful.

It can store data items of arbitrary size and alignment and makes no
assumptions about the type of the key used to locate entries.

Also included are some utility functions for hashing strings (or other
areas of memory) using the FNV-1a hash algorithm.