varnish: rework v3 and v4 support
By using VSC_Iter() to loop over the list of metrics exposed by varnish,
we can pick those we're interested *if they are found*.

Not explicitly referring to metrics in the VSC_C_main struct makes the
plugin more resilient to small differences between minor varnish
versions.

It also opens the possibility to monitor non-MAIN metrics, such as
per-backend or per-storage engine stats.

This patch should be compatible with the previous way of doing, from the
user point of view.

Fix #1302