Suppress successful putval responses to exec plugin
exec plugins are never supposed to receive responses
to putval commands. These responses are instead sent
to STDOUT, which is often attached to /dev/null.
Newer init systems, such as systemd, instead log
these responses.

This patch preserves the expected behavior by suppressing
the response if and only if the destination is STDOUT.