write_sensu: remove custom asprintf implementation
This was guarded by HAVE_ASPRINTF, but we never had a configure check
for asprintf (or vasprintf).

This only turned up when I made the function static, since they now
clash with the prototypes in stdio.h on FreeBSD.

As far as I know all the platforms we build on have these functions.
If this turns out to be not the case, we can always add compat functions
to our utilies later.

For now, this fixes the build on FreeBSD.