Revert "Turbostat: add missing types in configure.ac"
These autoconf tests were introduced in autoconf 2.60.
We still support older versions.

We've assumed for a while now that any platform we build
on has the types and functionality tested here available, and supports
a reasonable subset of the C99 and Posix standards.

This fixes building from git on RHEL5.

This reverts commit 45687863426b21f854711248258a13c345371843.