collectdmon: Add support for sending SIGKILL to collectd
Sometimes we might want to force termination of the collectd daemon,
but we do not have visibility over the PID of the collectd daemon. So
we need to request collectdmon to terminate the process for us, but
we cannot send SIGKILL directly as it's an uncatchable signal. Instead
we use SIGQUIT, catch that and then send SIGKILL to our child.