collectd.spec: disable plugins on older distro
Instead of keeping a list per RHEL version of which plugins to enable,
invert the logic and keep a list of plugins to disable when the distro
is too old. This way we only have to specify a plugin once, and this
also opens up the possibility to build on Fedora.