command parser: Add a vector-based interface.
The new interface expects a parsed list of fields. Parsing of input buffers
is now done centrally in the generic code and the existing parser is a
wrapper around this and the new vector-based parser. The actual command
parsers (PUTVAL for now) receive the parsed list of fields.

As a side-effect, this changes the parser behavior a little:

 - "foo=a string with spaces or \\ special characters" is now valid syntax
   for specifying options in addition to foo="value with spaces, etc.".
 - foo= is now a valid option even without quotes around the empty value.