memcached plugin: Use hostname_g when reading localhost.
Previously, the code would use the value of the "Host" option or
"127.0.0.1" if that option was unset (plus special cases for UNIX
sockets and a legacy mode). Obviously, "127.0.0.1" is a bad default.

This patch emulates the behavior of the MySQL plugin: if the "Host"
option is unset or set to either "localhost" or "127.0.0.1", the global
hostname_g variable is used.

Fixes: #801
Supersedes: #894