Switch to python-config for libpython detection
The old way of detecting libpython was unreliable and did not work on
OS X.

Instead use python-config which was introduced in python 2.6.

For older versions of python you can still set PYTHON_CPPFLAGS and
PYTHON_LDFLAGS on the ./configure command line.

By default we look for python3-config, python2-config and last
python-config. The path to python-config can be overriden by setting
PYTHON_CONFIG.

Fixes #888 and #27