varnish: drop support for version 2.x
Varnish 5 is due in a bit more than one month from now. Before adding
even more code & complexity to this plugin, remove the duplicate code
created in d010d9e and b208ddc.

For the record, the last commit to the 2.1 branch in varnish's git repo
was a bit more than 3 years ago, and the last release (2.1.5) was on
Jan 27, 2011.