virt plugin: Rewrite from scratch
This patch rewrites the existing virt plugin
almost entirely, borrowing only few bits from
the existing one.

The plugin exports as superset of the metrics the
old plugin exported, and accepts a superset of the
configuration settings the old plugin expected, so
it is intended to act as drop-in replacement.

The new plugin now depends on glib2, libvirt and libxml2.
The new plugin has tests, taking inspiration from the tests
of the ceph plugin (src/ceph_test.c).

The gathering of the libvirt values has been completely overhauled,
and now we use the libvirt bulk stats API:
https://libvirt.org/html/libvirt-libvirt-domain.html#virDomainListGetStats
which is a more modern and convenient way to fetch the
data from running libvirt domains.

We add resiliency in presence of unresponsive domains.
This can happen if a domain used shared storage (NFS, ISCSI)
and it experiences a network failure during one I/O operation.
check docs/README.virt.md to learn more about the issue
and the implemented solution.