chrony: fix type instances for unreachable sources
When a time source is unreachable, no sourcestats request is made and
the NaN values are pushed with an empty string as the type instance.
Modify chrony_request_source_data() and chrony_request_source_stats()
to share the address of the source, so the address is valid even when
sourcestats is not requested.