pcie_errors plugin: add syslog timestamp to event metadata
In case Source is set to 'logfile' add the time from syslog
event metadata. The event timestamp itself is added by collectd.
The new pattern is not mandatory because syslog format is
configurable and can be different on particular system.
Additionaly the default value for IsMandatory parameter is
changed to true as this is the most common use case.

Change-Id: Ic034e5f32a321bc83a62c877c903d053e4dea1eb
Signed-off-by: Kamil Wiatrowski <kamilx.wiatrowski@intel.com>