Push "collectd_plugin" to Lua registry to identify plugins.
Previously, lua_tostring() was used to identify readers/writers.
Unfortunately, it doesn't work on functions (it returns NULL).
Lua 5.2 and 5.3 come with luaL_tolstring() but it can (rather
theorethically in the current implementation) return identical
strings for two functions defined in two different Lua state.

This commit is a first step in adding unique identifiers to
readers and writers. It saves "lua/script.lua" in Lua registry
and uses it in register_read() and register_write() functions.

Unique identifier of read/write functions is still missing.